Split the User module in 3 modules
